DanielC | Invincibility card works
Player now will not take damage while the invincibility toggle is on. The timer is not in the player but rather the card. Can now cycle between cards correctly. Swapping to the next card will deactivate the current one

TODO - Add a card type enum to the base class so we know which overload to call from the player